// resources/lib/generated/api/types.gen.ts
export type InvitationAcceptPage = {
    allowPassword?: boolean;
    errorMessage?: string | null;
    invitation: TeamInvitationDetail;
    inviteeEmail?: string | null;
    isActivation?: boolean;
    isAuthenticated?: boolean;
    isCorrectUser?: boolean;
    isValid?: boolean;
    loginUrl?: string | null;
    oidcOptions?: Array<OidcProviderOption>;
    setPasswordUrl?: string | null;
};
// ...
export type OidcProviderOption = {
    label: string;
    provider: string;
    url: string;
};
// ...
export type TeamInvitationDetail = {
    expiresAt?: string | null;
    id: string;
    inviterEmail: string;
    inviterName: string;
    isExpired?: boolean;
    role: string;
    teamName: string;
    teamSlug: string;
};